backend/requirements.txt
fastapi
uvicornaiofiles
python-dotenv